Address PR review feedback
- Use GetDouble instead of GetFloat for Half reading (avoids extra conversion)
- Add Half scaffolding support to SqliteDatabaseModelFactory
  - Add HALF, FLOAT16 type hints for Half CLR type mapping
  - Add Half range validation in REAL value type handling
  - Add Half default value parsing

+++ src/Microsoft.Data.Sqlite.Core/SqliteValueReader.cs
@@ -233,7 +233,7 @@ internal abstract class SqliteValueReader
 #if NET6_0_OR_GREATER
         if (typeof(T) == typeof(Half))
         {
-            return (T)(object)(Half)GetFloat(ordinal);
+            return (T)(object)(Half)GetDouble(ordinal);
         }
 #endif
 
@@ -357,7 +357,7 @@ internal abstract class SqliteValueReader
 #if NET6_0_OR_GREATER
         if (type == typeof(Half))
         {
-            return (T)(object)(Half)GetFloat(ordinal);
+            return (T)(object)(Half)GetDouble(ordinal);
         }
 #endif
